requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile
#
certifi==2022.12.7
    # via requests
charset-normalizer==3.0.1
    # via requests
idna==3.4
    # via requests
pyyaml==6.0
    # via octanexchange (setup.py)
requests==2.28.2
    # via octanexchange (setup.py)
urllib3==1.26.14
    # via requests
